chore: move remaining mutations into Sell app
Move mutations still referenced from the Consign app into the Sell app.
