--- src/Apps/Consign/Routes/SubmissionFlow/Mutations/index.ts
@@ -1,4 +1,0 @@
-export * from "./addAssetToConsignment"
-export * from "./CreateConsignSubmissionMutation"
-export * from "./removeAssetFromConsignmentSubmission"
-export * from "./UpdateConsignSubmissionMutation"
