+++ src/Apps/Sell/Utils/createOrUpdateConsignSubmission.ts
@@ -1,7 +1,7 @@
 import {
   createConsignSubmissionMutation,
   updateConsignSubmissionMutation,
-} from "Apps/Consign/Routes/SubmissionFlow/Mutations"
+} from "Apps/Sell/Mutations"
 import { CreateSubmissionMutationInput } from "__generated__/CreateConsignSubmissionMutation.graphql"
 import { UpdateSubmissionMutationInput } from "__generated__/UpdateConsignSubmissionMutation.graphql"
 import { Environment } from "react-relay"
